+++ examples/acme-cli.ts
@@ -6,6 +6,7 @@ import {
   DnsUtils,
   TimeoutError,
 } from "jsr:@fishballpkg/acme";
+import { resolveDns } from "jsr:@fishballpkg/acme/resolveDns.deno";
 
 const alertList = (...messages: string[]) => {
   for (const message of messages) {
@@ -88,7 +89,7 @@ while (true) {
   console.log("⏳ Polling DNS to verify TXT records are updated correctly...");
   try {
     await Promise.all(expectedRecords.map(async ({ name, content }) => {
-      await DnsUtils.pollDnsTxtRecord(name, { pollUntil: content });
+      await DnsUtils.pollDnsTxtRecord(name, { pollUntil: content, resolveDns });
     }));
     break;
   } catch (e) {