+++ src/DnsUtils/mod.ts
@@ -4,6 +4,6 @@
  * Utility functions to help you with `dns-01` challenge
  */
 
-export * from "./findAuthoritativeNameServerIps.ts";
 export * from "./pollDnsTxtRecord.ts";
 export * from "./resolveDns.ts";
+export * from "./createUnanimousResolveDns.ts";