+++ src/DnsUtils/resolveDns.deno.ts
@@ -1,1 +1,0 @@
-export const resolveDns = Deno.resolveDns;